requirements_test.txt
moto[all]==2.0.0
freezegun==1.1.0
pytest-asyncio==0.14.0
